+++ eslint.config.cjs
@@ -1,0 +1,29 @@
+const tsParser = require('@typescript-eslint/parser');
+const tsPlugin = require('@typescript-eslint/eslint-plugin');
+
+module.exports = [
+    {
+        ignores: ['built/**', 'node_modules/**'],
+    },
+    {
+        files: ['src/**/*.ts'],
+        languageOptions: {
+            parser: tsParser,
+            ecmaVersion: 'latest',
+            sourceType: 'module',
+        },
+        plugins: {
+            '@typescript-eslint': tsPlugin,
+        },
+        rules: {
+            quotes: [
+                'error',
+                'single',
+                {
+                    avoidEscape: false,
+                    allowTemplateLiterals: true,
+                },
+            ],
+        },
+    },
+];